Fully functional VC++ 7.1 solution
git-svn-id: http://eathena.googlecode.com/svn/branches/eapp@8111 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/map/chrif.cpp
@@ -1858,7 +1858,7 @@ int chrif_parse_var_recv(int fd)
 		set_var((const char *)var.name(), 		
 			(postfix=='$') ?
 			((void *)(const char*)var.value()) :
-			((void *)(ssize_t)atol(var.value())) );
+			((void *)(size_t)(ssize_t)atol(var.value())) );
 
 	}
 	return 0;
